Fix padding and hidden scrollbar
Rewrote the View structure for page_view and ui_view to pad between the
TextView with the page content and the scrollbar. So now there's inner
and outer padding basically with the inner padding on the content making
up for the 0 right padding on the ui_view.

Also removed some unused "use" statements.